warrior-install.sh: use pip3 if available
The standard Warrior Docker image comes with both Python2 and Python3,
with separate pip and pip3 binaries, respectively. Because it runs the
Warrior using Python3, installing warcio using pip in warrior-install.sh
will not make it available to the grabber. This change fixes this by
using pip3 if it is available.